requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
bcrypt==3.2.2
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
datefinder==0.7.1
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-allauth==0.51.0
django-bootstrap3==21.2
django-braces==1.15.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
misaka==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.1.1
protobuf==3.20.1
psycopg2 @ file:///C:/ci_310/psycopg2_1642085998310/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==4.1.1
urllib3==1.26.9
wincertstore==0.2
